Dismiss the gallery if restoring after process death
Gallery items will not be available after process death, and a user will see "Empty gallery" message even if there are gallery items in a thread. We cannot save gallery items to the bundle in onSaveInstanceState and restore them after process death because if there are too many gallery items, saving them to the bundle will lead to TransactionTooLargeException.

So, i think this is better to dismiss the gallery after process death for better user experience until the proper restoration of the gallery state exists.