Show more informative error message in the gallery when cached media file cannot be obtained
Some users have reported an issue with the cache resulting in an error message that reads "cache is not available" in the gallery. However, I have not been able to replicate this issue on my own devices, and the current error message is too vague to determine the cause of the problem. To address this, I have created a new method that will obtain a cached media file and throw an exception with a more informative message if the cache media directory cannot be obtained. This new message will be displayed instead of the current "cache is not available" message.